% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Haver data on US GDP and contributions.}
\format{
\subsection{\code{gdp}}{

A data frame with quarterly rows and 7 columns for GDP, contributions and dates.
}
}
\source{
Haver
}
\usage{
gdp
}
\description{
Data from Haver on quarterly US GDP and its contributions.
}
\keyword{datasets}
