% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifs}
\alias{ifs}
\title{IMF IFS data on exchange rates}
\format{
\subsection{\code{ifs}}{

A data frame with annual rows and 6 columns for 5 countries and dates.
}
}
\source{
\url{https://datahelp.imf.org/knowledgebase/articles/838041-sdmx-2-0-restful-web-service}
}
\usage{
ifs
}
\description{
Data from the IMF IFS dataset available at http://dataservices.imf.org
}
\keyword{datasets}
