% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr-package.R
\docType{package}
\name{imfbookr-package}
\alias{imfbookr}
\alias{imfbookr-package}
\title{imfbookr: Generate IMF-style Charts}
\description{
Library to generate IMF-style charts for publications. Functions to generate line charts, bar charts, scatter charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://r-tutorial}
}

}
\author{
\strong{Maintainer}: Bas Bakker \email{BBakker@imf.org}

Authors:
\itemize{
  \item Greta Polo \email{gpolo@imf.org}
}

}
\keyword{internal}
