% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{imfdata_by_countries_and_series}
\alias{imfdata_by_countries_and_series}
\title{Retrieve IMF Data by Countries and Series}
\usage{
imfdata_by_countries_and_series(
  department,
  dataset,
  countries = NULL,
  series = NULL,
  frequency = "A",
  needs_auth = FALSE,
  needs_labels = FALSE
)
}
\arguments{
\item{department}{A character string representing the department.}

\item{dataset}{A character string representing the dataset.}

\item{countries}{A vector of country codes (optional).}

\item{series}{A vector of series codes (optional).}

\item{frequency}{A character string representing the frequency ("A", "Q", or "M").
The default is "A".}

\item{needs_auth}{Logical; whether authentication is required. Default is \code{FALSE}.}

\item{needs_labels}{Logical; whether to retrieve data with labels. Default is \code{FALSE}.}
}
\value{
A data frame containing the retrieved data.
}
\description{
This function retrieves data from the IMF database using the specified
countries and series. It builds the necessary key and calls an internal function
to handle data retrieval and processing.
}
\examples{
\dontrun{
  # Retrieve annual data for a set of countries and series
  df <- imfdata_by_countries_and_series(department = "DEPT", dataset = "DATASET",
                                         countries = c("US", "GB"), series = c("SERIES1", "SERIES2"),
                                         frequency = "A")
}

}
