% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{imfdata_by_key}
\alias{imfdata_by_key}
\title{Retrieve IMF Data by Key}
\usage{
imfdata_by_key(
  department,
  dataset,
  key,
  needs_auth = FALSE,
  needs_labels = FALSE
)
}
\arguments{
\item{department}{A character string representing the department.}

\item{dataset}{A character string representing the dataset.}

\item{key}{A character string or list containing keys for countries, series, and frequency.}

\item{needs_auth}{Logical; whether authentication is required.}

\item{needs_labels}{Logical; whether to retrieve data with labels.}
}
\value{
A data frame with the retrieved and processed data.
}
\description{
This function retrieves IMF data using the provided key.
It handles authentication (if required), builds the API call, and processes
the returned SDMX data.
}
