% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{imfdata_get_attributes_dataset}
\alias{imfdata_get_attributes_dataset}
\title{Get Attributes of an IMF Dataset}
\usage{
imfdata_get_attributes_dataset(dataset)
}
\arguments{
\item{dataset}{A character string specifying the dataset identifier.}
}
\value{
A data frame with one column, \code{attributes}, listing the attribute names of the dataset.
}
\description{
This function retrieves the data structure of a given IMF dataset and extracts its attributes.
}
\examples{
\dontrun{
  attrs <- imfdata_get_attributes_dataset("DSD_CPI")
  print(attrs)
}

}
