% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{imfdata_get_dimensions_dataset}
\alias{imfdata_get_dimensions_dataset}
\title{Get Dimensions of an IMF Dataset}
\usage{
imfdata_get_dimensions_dataset(dataset)
}
\arguments{
\item{dataset}{A character string specifying the dataset identifier.}
}
\value{
A data frame with one column, \code{Dimensions}, listing the dimension names of the dataset.
}
\description{
This function retrieves the data structure of a given IMF dataset and extracts its dimensions.
}
\examples{
\dontrun{
  dims <- imfdata_get_dimensions_dataset("DSD_CPI")
  print(dims)
}

}
