% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{imfdata_show_datasets}
\alias{imfdata_show_datasets}
\title{Show Available IMF Datasets}
\usage{
imfdata_show_datasets(needs_auth = FALSE)
}
\arguments{
\item{needs_auth}{Logical; if \code{TRUE} the function will use authentication. Default is \code{FALSE}.}
}
\value{
A data frame of available IMF datasets.
}
\description{
This function retrieves a data frame containing information on available IMF datasets by querying the IMF API.
If authentication is required, the function obtains an access token via Azure AD B2C.
}
\examples{
\dontrun{
  datasets <- imfdata_show_datasets(needs_auth = TRUE)
  head(datasets)
}

}
