% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{processTimePeriod}
\alias{processTimePeriod}
\title{Process the TIME_PERIOD Column in a Dataset}
\usage{
processTimePeriod(dataset, frequency)
}
\arguments{
\item{dataset}{A data frame that includes a \code{TIME_PERIOD} column.}

\item{frequency}{A character vector of length 1 indicating the frequency ("Q", "M", or "A").}
}
\value{
The dataset with a new \code{date} column computed from \code{TIME_PERIOD}.
}
\description{
This function processes the \code{TIME_PERIOD} column in a dataset based on the provided frequency.
}
\examples{
\dontrun{
  df <- data.frame(TIME_PERIOD = c("2021-05", "2021-06"))
  processTimePeriod(df, "M")
}

}
