% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{yearMonthToDate}
\alias{yearMonthToDate}
\title{Convert a Year-Month String to a Date}
\usage{
yearMonthToDate(month_string)
}
\arguments{
\item{month_string}{A character string in the format "YYYY-MM" (or similar).}
}
\value{
A \code{yearmon} object representing the last day of the month.
}
\description{
This function converts a string representing a year and month to a date
corresponding to the last day of that month.
}
\examples{
yearMonthToDate("2021-05")

}
