% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata-wrapper.R
\name{yearQuarterToDate}
\alias{yearQuarterToDate}
\title{Convert a Year-Quarter String to a Date}
\usage{
yearQuarterToDate(quarter_string)
}
\arguments{
\item{quarter_string}{A character string in the format "YYYY-Q" where Q is the quarter number.}
}
\value{
A \code{yearqtr} object representing the quarter.
}
\description{
This function converts a string representing a year and quarter to a date
corresponding to the last month of that quarter.
}
\examples{
yearQuarterToDate("2021-Q1")

}
